Fixed bugs + updated product record form
Fixed an issue where the library button controls would throw an exception and would be rendered useless for the lifetime of the application.

It is to be noted that InvokeRequired will return false if the control handle has not been created.

Also, the product record form has been updated to appopriately show what information the database has. It is currently read-only.

- Updated folders attribute when parsing from the database entry to an extraction record.
- Added global main thread ID to determine if a thread is the main (UI) thread.
- Updated design and functionality of Product Record Form.
- PageButtonControl.UpdateControl() now checks if it is invokable including if the handle is created.