[Not complete] Revamp structure
Instead of the interface IDPWorkingFile, all files that are discovered from archives are of base type DPAbstractFile.

DPFile is still the same and inherits DPAbstractFile

DPAbstractFile will be the abstract class for archive files which inherits DPAbstractFile.

DPRARFile inherits from DPAbstractArchive and same for 7z and Zip

Added a few structures and enums as well.

Removed and revamped some code for performance or memory improvements.