Fix weird bugs with versioning + added no-version arg
There was a weird bug where using async would not complete properly - as in, one file would be updated but the others aren't even updated.

Furthermore, when switched to sync, there was encoding issues, that is human readable but deemed corrupted to dotnet. It was attempting use another encoder (Western something).

Also added a no-version argument to skip incrementing the argument.