Fixed cancel button not calling CancelJob
This one was a tricky one to find. In short, because we use BeginInvoke, the CancellationTokenSource was not immediately being updated. As a result, StartProgress immediately exits and for DPExtractJob, it would call Token, getting a stale token. Then, the StartProgress code woudl update the token source. Therefore, any cancellation calls would go to the new cancellation source which did not have any registrations.